+++ components/chainhook-sdk/src/chainhooks/types/stacks.rs
@@ -3,7 +3,10 @@ use std::collections::BTreeMap;
 use chainhook_types::StacksNetwork;
 use schemars::JsonSchema;
 
-use super::{BlockIdentifierIndexRule, ChainhookSpecification, ExactMatchingRule, HookAction};
+use super::{
+    BlockIdentifierIndexRule, ChainhookConfig, ChainhookSpecification, ExactMatchingRule,
+    HookAction,
+};
 
 #[derive(Clone, Debug, Serialize, Deserialize, PartialEq, JsonSchema)]
 pub struct StacksChainhookFullSpecification {
@@ -181,4 +184,34 @@ pub struct StacksNftEventBasedPredicate {
 #[serde(rename_all = "snake_case")]
 pub struct StacksStxEventBasedPredicate {
     pub actions: Vec<String>,
+}
+
+impl ChainhookConfig {
+    pub fn deregister_stacks_hook(
+        &mut self,
+        hook_uuid: String,
+    ) -> Option<StacksChainhookSpecification> {
+        let mut i = 0;
+        while i < self.stacks_chainhooks.len() {
+            if self.stacks_chainhooks[i].uuid == hook_uuid {
+                let hook = self.stacks_chainhooks.remove(i);
+                return Some(hook);
+            } else {
+                i += 1;
+            }
+        }
+        None
+    }
+
+    pub fn expire_stacks_hook(&mut self, hook_uuid: String, block_height: u64) {
+        let mut i = 0;
+        while i < self.stacks_chainhooks.len() {
+            if ChainhookSpecification::stacks_key(&self.stacks_chainhooks[i].uuid) == hook_uuid {
+                self.stacks_chainhooks[i].expired_at = Some(block_height);
+                break;
+            } else {
+                i += 1;
+            }
+        }
+    }
 }
